refactor : 스크래핑 코드에서 data get 하는 부분 분리
